constrained_sampling/requirements.txt
absl-py==2.1.0
accelerate==0.29.3
adan-pytorch==0.1.0
aiohttp==3.9.3
aiosignal==1.3.1
aitextgen==0.6.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
attrs==23.2.0
bash_kernel==0.9.3
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
diffusers==0.27.2
einops==0.7.0
executing==2.0.1
filelock==3.13.3
fire==0.6.0
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gdown==5.1.0
grpcio==1.62.1
huggingface-hub==0.22.0
hydra-core==1.3.2
idna==3.6
imhist==0.0.4
importlib_metadata==7.1.0
imWatermark==0.0.2
invisible-watermark==0.2.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lightning-utilities==0.11.1
lmdb==1.4.1
lpips==0.1.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open-clip-torch==2.24.0
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==3.20.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyDeprecate==0.3.2
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.1
pytz==2024.1
PyWavelets==1.5.0
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scipy==1.12.0
sentencepiece==0.2.0
setuptools==68.2.2
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
-e git+https://github.com/CompVis/taming-transformers@3ba01b241669f5ade541ce990f7650a3b8f65318#egg=taming_transformers
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
timm==0.9.16
tokenizers==0.15.2
torch==2.2.1
torch-fidelity==0.3.0
torchmetrics==1.3.2
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.41.2
yarl==1.9.4
zipp==3.18.1
